refactor(print_statement): a function can print one line bpf code from statement
It can be used in emulate and disasm.
It print with color, so emulate need to set the color itself, otherwise
print_statement might print the color emulate doesn't want.
But we still need a function to extract statement from raw bpf_code and
do some examination. Because I DID ZERO CHECK in print_statement!
When emulate call print_statement, the statement should be checked by
resolver, so it's fine